refactor(preact): share component-seam helpers with react adapter
Extract the host-internal field interception and the capitalized
subcomponent wrapper into a framework-agnostic seam module, consumed by
both adapters via @expressive/react/state. Removes the duplicated loops
from both component.ts; preact retains only its render-restore step.
