+++ packages/react/src/state.ts
@@ -3,6 +3,7 @@ import { Runtime, useFactory, useHook, useReady } from './runtime';
 
 export { Runtime, useFactory, useHook, useReady };
 export { Consumer, Provider, Context, Layers } from './context';
+export { intercept, defineSubcomponent } from './seam';
 export { use } from './use';
 
 /** Type may not be undefined - instead will be null.  */
